feat(core): multi-edit chain via update(Telescope, fn) + extract conversion builders to sibling files
- Add chain accumulator on Telescope: .with(fn) / .apply(S) and .update(Telescope, fn) for pre-built path reuse
- Java overload resolution disambiguates .update(S, fn) (terminal) vs .update(Telescope, fn) (chainable)
- Rename .field(String) → .fieldByName(String) and .field(String, Class<B>) → .fieldByName(String, Class<B>) to signal runtime-checked nature
- Extract conversion builders (From, To, BeanFrom, BeanTo, MapBeanFrom, MapBeanTo, MapTo, MapBuilder, FieldMapping, Mapper) from nested Telescope.* into sibling files in the same package
- Bump version references to 0.3.0
- Update README + CLAUDE.md with multi-edit chain documentation and compile-safety scoring
- Address Copilot PR review feedback (String::trim in README, auto() javadoc clarification, drop unused sourceType from Mapper)
